Remove dependency on 'assists' from 'completion' crate
